refactor: Delete post and associated comments by ID
This commit refactors the `deletePostById` method in the `PostController` class to delete a post and its associated comments by ID. The `deletePostById` method now calls a new method `deleteUsernameInComments` to delete the comments related to the post being deleted. This ensures that all comments associated with the post are also removed from the database.

The `deleteUsernameInComments` method is also added to the `UserController` class to delete comments made by a specific user. This method is called when a user deletes their account.

These changes improve the data integrity and ensure that all related data is properly deleted when a post or user is deleted.